@@ -3,45 +3,45 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^5.9.0",
     "@date-io/date-fns": "^1.3.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-brands-svg-icons": "^6.2.1",
+    "@fortawesome/free-regular-svg-icons": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/pickers": "^3.3.10",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/lodash": "^4.14.172",
-    "@types/node": "^12.20.19",
-    "@types/react": "^17.0.17",
-    "@types/react-dom": "^17.0.9",
-    "axios": "^0.21.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.5",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "axios": "^1.2.2",
     "date-fns": "^2.23.0",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-loading-skeleton": "^2.2.0",
-    "react-moment": "^1.1.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "react-select": "^4.3.1",
-    "react-slick": "^0.28.1",
-    "react-toastify": "^8.0.2",
+    "moment": "^2.29.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-loading-skeleton": "^3.1.0",
+    "react-moment": "^1.1.2",
+    "react-responsive": "^9.0.2",
+    "react-router-dom": "^6.6.1",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.7.0",
+    "react-slick": "^0.29.0",
+    "react-toastify": "^9.1.1",
     "slick-carousel": "^1.8.1",
-    "swr": "^1.3.0",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.1.2"
+    "swr": "^2.0.0",
+    "typescript": "^4.9.4",
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
-    "start": "cross-env TAILWIND_MODE=watch craco start",
-    "build": "cross-env TAILWIND_MODE=build craco build",
+    "start": "cross-env TAILWIND_MODE=watch react-scripts start",
+    "build": "cross-env TAILWIND_MODE=build react-scripts build",
     "serve": "npx serve -l 3000 -n -s",
-    "test": "craco test",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -63,21 +63,23 @@
     ]
   },
   "devDependencies": {
-    "@tailwindcss/aspect-ratio": "^0.2.1",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
     "@types/date-fns": "^2.6.0",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/react-select": "^4.0.17",
-    "@types/react-slick": "^0.23.5",
-    "@typescript-eslint/eslint-plugin": "^4.29.1",
-    "@typescript-eslint/parser": "^4.29.1",
-    "autoprefixer": "^9.8.6",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-select": "^5.0.0",
+    "@types/react-slick": "^0.23.10",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/parser": "^5.47.1",
+    "autoprefixer": "^10.4.13",
     "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "postcss": "^7.0.36",
-    "prettier": "^2.3.2",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.7"
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "os-browserify": "^0.3.0",
+    "path-browserify": "^1.0.1",
+    "postcss": "^8.4.20",
+    "prettier": "^2.8.1",
+    "tailwindcss": "^3.2.4"
   },
   "optionalDependencies": {
     "fsevents": "2.1.2"
