@@ -1,49 +1,37 @@
-import { Dispatch, SetStateAction, useEffect } from 'react';
-import { useRef, useState } from 'react';
-import { FC } from 'react';
-import { Link, useHistory, useLocation } from 'react-router-dom';
-import { useGetSWR } from '../../../hooks/api';
-import { IResourceContent } from '../../../interfaces';
-import Ellipsis from './Ellipsis';
+import * as React from 'react';
+import { Link, useNavigate, useLocation } from 'react-router-dom';
+import useSWR from 'swr';
+import useImmutableSWR from 'swr/immutable';
+import { IInfoMe, IResourceContent } from '../../../interfaces';
+import ComponentEllipsis from './Ellipsis';
+import { axiosGetfetcher } from '../../../hooks/api';
+import TokenContext from '../../../store/TokenContext';
 
-interface HeaderProps {
-  token: string | null;
-  setToken: Dispatch<SetStateAction<string | null>>;
-  setRememberToken: (
-    value: string | ((val: string | null) => string | null) | null,
-  ) => void;
-  nickname: string | null;
-  userType: string | null;
-  setUserType: Dispatch<SetStateAction<string | null>>;
-  setUserNickname: Dispatch<SetStateAction<string | null>>;
-}
+interface IPropsComponentHeader {}
 
-const Header: FC<HeaderProps> = ({
-  token,
-  setToken,
-  setRememberToken,
-  nickname,
-  userType,
-  setUserType,
-  setUserNickname,
-}) => {
-  const history = useHistory();
+const ComponentHeader: React.FC<IPropsComponentHeader> = ({}) => {
+  const navigate = useNavigate();
   const location = useLocation();
-  const { data: headerLogo } = useGetSWR<IResourceContent[]>(
-    `${process.env.REACT_APP_BACK_URL}/resource/header_logo`,
-    null,
-    false,
+  const tokenCtx = React.useContext(TokenContext);
+  const { token, setToken } = tokenCtx;
+  const [isVisibleEllipsis, setIsVisibleEllipsis] =
+    React.useState<boolean>(false);
+  const [isVisibleMenu, setIsVisibleMenu] = React.useState<boolean>(false);
+  const menuElementRef = React.useRef<HTMLDivElement | null>(null);
+  const { data: dataGetMe, error: errorGetMe } = useSWR<IInfoMe>(
+    !!token ? `${process.env.REACT_APP_BACK_URL}/user/me` : null,
+    () => axiosGetfetcher(`${process.env.REACT_APP_BACK_URL}/user/me`, token),
+    { revalidateOnFocus: false, revalidateIfStale: false },
   );
-  const logoutHandler = () => {
-    setToken('');
-    localStorage.setItem('token', '');
-    setUserType('');
-    setUserNickname('');
-    history.go(0);
-  };
-  const [isVisibleEllipsis, setIsVisibleEllipsis] = useState<boolean>(false);
-  const [isVisibleMenu, setIsVisibleMenu] = useState<boolean>(false);
-  const menuElementRef = useRef<HTMLDivElement | null>(null);
+  const { data: dataHeaderLogo } = useImmutableSWR<IResourceContent[]>(
+    `${process.env.REACT_APP_BACK_URL}/resource/header_logo`,
+    () =>
+      axiosGetfetcher(
+        `${process.env.REACT_APP_BACK_URL}/resource/header_logo`,
+        token,
+      ),
+    { revalidateOnFocus: false, revalidateIfStale: false },
+  );
   const menuPositionHandler = (event: any) => {
     if (
       isVisibleMenu &&
@@ -52,7 +40,7 @@ const Header: FC<HeaderProps> = ({
       setIsVisibleMenu(false);
     }
   };
-  const ellipsisElementRef = useRef<HTMLDivElement | null>(null);
+  const ellipsisElementRef = React.useRef<HTMLDivElement | null>(null);
   const ellipsisPositionHandler = (event: any) => {
     if (
       isVisibleEllipsis &&
@@ -62,11 +50,15 @@ const Header: FC<HeaderProps> = ({
       setIsVisibleEllipsis(false);
     }
   };
-  useEffect(() => {
+  const logoutHandler = () => {
+    setToken('');
+    navigate(0);
+  };
+  React.useEffect(() => {
     window.addEventListener('click', menuPositionHandler);
     return () => window.removeEventListener('click', menuPositionHandler);
   }, [isVisibleMenu]);
-  useEffect(() => {
+  React.useEffect(() => {
     window.addEventListener('click', ellipsisPositionHandler);
     return () => window.removeEventListener('click', ellipsisPositionHandler);
   }, [isVisibleEllipsis]);
@@ -74,20 +66,23 @@ const Header: FC<HeaderProps> = ({
     <div className="h-[100px] font-semibold text-gray-300 font-noto header-container">
       <div className="xl:max-w-[1152px] lg:max-w-[864px] md:max-w-[680px] md:flex hidden h-full mx-auto justify-center items-center">
         <div className="flex items-center">
-          <Link to="/">
-            {headerLogo && headerLogo.length > 0 ? (
+          {!!dataHeaderLogo && dataHeaderLogo.length > 0 && (
+            <Link to="/">
               <img
                 className="xl:mr-[80px] lg:mr-[47px] md:mr-[40px] mr-[20px]"
-                src={headerLogo[0].content}
+                src={dataHeaderLogo[0].content}
                 width="110.24"
                 alt="Logo"
               />
-            ) : (
-              ''
-            )}
-          </Link>
+            </Link>
+          )}
           <div className="flex flex-nowrap">
-            {((token && userType !== 'admin') || token === '' || !token) && (
+            {!(
+              !!token &&
+              !!dataGetMe &&
+              !!!errorGetMe &&
+              dataGetMe.role === 'admin'
+            ) && (
               <>
                 <Link to="/">
                   <button
@@ -110,146 +105,156 @@ const Header: FC<HeaderProps> = ({
           </div>
         </div>
         <div>
-          {token && nickname && userType !== 'admin' && (
-            <div
-              className="xl:ml-[766px] lg:ml-[512px] md:ml-[334px]"
-              onMouseEnter={() => setIsVisibleEllipsis(true)}
-              onMouseLeave={() => setIsVisibleEllipsis(false)}
-            >
-              <button className="rounded-full w-[40px] h-[40px] flex items-center justify-center bg-[#8dc556] leading-[150%] text-white font-semibold">
-                {nickname.charAt(0)}
-              </button>
-              {isVisibleEllipsis && (
-                <div className="fixed z-[999] min-w-max">
-                  <Ellipsis
-                    token={token}
-                    setToken={setToken}
-                    logoutHandler={logoutHandler}
-                  />
-                </div>
-              )}
-            </div>
-          )}
-          {(token === '' || !token) && (
+          {!!token &&
+            !!dataGetMe &&
+            !!!errorGetMe &&
+            !!dataGetMe.nickname &&
+            dataGetMe.role === 'student' && (
+              <div
+                className="xl:ml-[766px] lg:ml-[512px] md:ml-[334px]"
+                onMouseEnter={() => setIsVisibleEllipsis(true)}
+                onMouseLeave={() => setIsVisibleEllipsis(false)}
+              >
+                <button className="rounded-full w-[40px] h-[40px] flex items-center justify-center bg-[#8dc556] leading-[150%] text-white font-semibold">
+                  {dataGetMe.nickname.charAt(0)}
+                </button>
+                {isVisibleEllipsis && (
+                  <div className="fixed z-[999] min-w-max">
+                    <ComponentEllipsis logoutHandler={logoutHandler} />
+                  </div>
+                )}
+              </div>
+            )}
+          {!(!!token && !!dataGetMe && !!!errorGetMe) && (
             <div className="xl:ml-[586px] lg:ml-[360px] md:ml-[230px]">
               <Link to="/signin">
-                <button className="bg-white font-[14px] leading-[21px] font-semibold text-[#808695] border-[1px] border-[#DCDEE2] box-border rounded-[40px] h-[41px] xl:w-[99px] lg:w-[84.48px] md:w-[63.36px]">
+                <button className="bg-white text-[14px] leading-[21px] font-semibold text-[#808695] border-[1px] border-[#DCDEE2] box-border rounded-[40px] h-[41px] xl:w-[99px] lg:w-[84.48px] md:w-[63.36px]">
                   로그인
                 </button>
               </Link>
               <Link to="/signup">
-                <button className="ml-[12px] font-[14px] leading-[21px] font-semibold text-white bg-[#8DC556] rounded-[40px] h-[41px] xl:w-[112px] lg:w-[95.57px] md:w-[71.68px]">
+                <button className="ml-[12px] text-[14px] leading-[21px] font-semibold text-white bg-[#8DC556] rounded-[40px] h-[41px] xl:w-[112px] lg:w-[95.57px] md:w-[71.68px]">
                   회원가입
                 </button>
               </Link>
             </div>
           )}
-          {token && nickname && userType === 'admin' && (
-            <button
-              className="flex items-center justify-center xl:ml-[800px] lg:ml-[550px] md:ml-[400px]"
-              onClick={logoutHandler}
-            >
-              로그아웃
-            </button>
-          )}
+          {!!token &&
+            !!dataGetMe &&
+            !!!errorGetMe &&
+            dataGetMe.role === 'admin' && (
+              <button
+                type="button"
+                className="flex justify-center items-center w-max xl:ml-[800px] lg:ml-[550px] md:ml-[400px]"
+                onClick={logoutHandler}
+              >
+                로그아웃
+              </button>
+            )}
         </div>
       </div>
       <div className="flex items-center justify-between h-full mx-auto md:hidden">
         <div className="flex items-center justify-center w-full">
           <Link to="/">
-            {headerLogo && headerLogo.length > 0 ? (
-              <img src={headerLogo[0].content} width="110.24" alt="Logo" />
+            {!!dataHeaderLogo && dataHeaderLogo.length > 0 ? (
+              <img src={dataHeaderLogo[0].content} width="110.24" alt="Logo" />
             ) : (
               ''
             )}
           </Link>
-          {((token && userType !== 'admin') || token === '' || !token) && (
-            <>
-              <button
-                className={`rounded-[1px] min-w-max ${
-                  token === '' || !token
-                    ? 'sm:mx-[112px] xs:mx-[63px]'
-                    : token && nickname && userType !== 'admin'
-                    ? 'sm:mx-[165px] xs:mx-[115px]'
-                    : ''
-                }`}
-                onClick={() => setIsVisibleMenu(!isVisibleMenu)}
+          <>
+            <button
+              className={`rounded-[1px] min-w-max ${
+                !!!token
+                  ? 'sm:mx-[112px] xs:mx-[63px]'
+                  : !!token &&
+                    !!dataGetMe &&
+                    !!!errorGetMe &&
+                    dataGetMe.role === 'student'
+                  ? 'sm:mx-[165px] xs:mx-[115px]'
+                  : ''
+              }`}
+              onClick={() => setIsVisibleMenu(!isVisibleMenu)}
+            >
+              메뉴
+            </button>
+            {isVisibleMenu && (
+              <div
+                ref={menuElementRef}
+                className="flex-none border-[1px] border-[#DCDEE2] box-border rounded-[10px] bg-white min-w-full absolute top-[100px] z-[999]"
               >
-                메뉴
-              </button>
-              {isVisibleMenu && (
-                <div
-                  ref={menuElementRef}
-                  className="flex-none border-[1px] border-[#DCDEE2] box-border rounded-[10px] bg-white min-w-full absolute top-[100px] z-[999]"
+                <Link to="/">
+                  <button
+                    className={`block w-full px-[10px] py-[10px] ${
+                      location.pathname === '/'
+                        ? 'text-[#8DC556]'
+                        : 'text-[#515A6E]'
+                    }`}
+                  >
+                    강의
+                  </button>
+                </Link>
+                <a href="mailto:mpnaroo@naver.com">
+                  <button className="block w-full px-[10px] py-[10px]">
+                    문의하기
+                  </button>
+                </a>
+              </div>
+            )}
+          </>
+
+          {!!token &&
+            !!dataGetMe &&
+            !!!errorGetMe &&
+            !!dataGetMe.nickname &&
+            dataGetMe.role === 'student' && (
+              <div>
+                <button
+                  className="rounded-full w-[40px] h-[40px] flex items-center justify-center bg-[#8dc556] leading-[150%] text-white font-semibold"
+                  onClick={() => setIsVisibleEllipsis(!isVisibleEllipsis)}
                 >
-                  <Link to="/">
-                    <button
-                      className={`block w-full px-[10px] py-[10px] ${
-                        location.pathname === '/'
-                          ? 'text-[#8DC556]'
-                          : 'text-[#515A6E]'
-                      }`}
-                    >
-                      강의
-                    </button>
-                  </Link>
-                  <a href="mailto:mpnaroo@naver.com">
-                    <button className="block w-full px-[10px] py-[10px]">
-                      문의하기
-                    </button>
-                  </a>
-                </div>
-              )}
-            </>
-          )}
-          {token && nickname && userType !== 'admin' && (
-            <div>
-              <button
-                className="rounded-full w-[40px] h-[40px] flex items-center justify-center bg-[#8dc556] leading-[150%] text-white font-semibold"
-                onClick={() => setIsVisibleEllipsis(!isVisibleEllipsis)}
-              >
-                {nickname.charAt(0)}
-              </button>
-              {isVisibleEllipsis && (
-                <div
-                  ref={ellipsisElementRef}
-                  className="relative z-[999] min-w-max"
-                >
-                  <Ellipsis
-                    token={token}
-                    setToken={setToken}
-                    logoutHandler={logoutHandler}
-                  />
-                </div>
-              )}
-            </div>
-          )}
+                  {dataGetMe.nickname.charAt(0)}
+                </button>
+                {isVisibleEllipsis && (
+                  <div
+                    ref={ellipsisElementRef}
+                    className="relative z-[999] min-w-max"
+                  >
+                    <ComponentEllipsis logoutHandler={logoutHandler} />
+                  </div>
+                )}
+              </div>
+            )}
           {(token === '' || !token) && (
             <div>
               <Link to="/signin">
-                <button className="bg-white font-[10px] leading-[10px] font-semibold text-[#808695] border-[1px] border-[#DCDEE2] box-border rounded-[40px] w-[60px] h-[41px]">
+                <button className="bg-white text-[10px] leading-[10px] font-semibold text-[#808695] border-[1px] border-[#DCDEE2] box-border rounded-[40px] w-[60px] h-[41px]">
                   로그인
                 </button>
               </Link>
               <Link to="/signup">
-                <button className="ml-[12px] font-[10px] leading-[10px] font-semibold text-white bg-[#8DC556] rounded-[40px] w-[70px] h-[41px]">
+                <button className="ml-[12px] text-[10px] leading-[10px] font-semibold text-white bg-[#8DC556] rounded-[40px] w-[70px] h-[41px]">
                   회원가입
                 </button>
               </Link>
             </div>
           )}
-          {token && nickname && userType === 'admin' && (
-            <button
-              className="flex items-center justify-center ml-[50%]"
-              onClick={logoutHandler}
-            >
-              로그아웃
-            </button>
-          )}
+          {!!token &&
+            !!dataGetMe &&
+            !!!errorGetMe &&
+            !!dataGetMe.nickname &&
+            dataGetMe.role === 'admin' && (
+              <button
+                className="flex items-center justify-center ml-[50%]"
+                onClick={logoutHandler}
+              >
+                로그아웃
+              </button>
+            )}
         </div>
       </div>
     </div>
   );
 };
 
-export default Header;
+export default ComponentHeader;
