@@ -1,8 +1,10 @@
-import ReactDOM from 'react-dom';
+import * as ReactDOM from 'react-dom/client';
 import './index.css';
 
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import AppRouterWrapper from './App';
 
-ReactDOM.render(<AppRouterWrapper />, document.getElementById('root'));
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+  <AppRouterWrapper />,
+);
