@@ -1,28 +1,34 @@
 import { FC } from 'react';
-import { DataResponse } from '../../hooks/api';
+import { KeyedMutator } from 'swr';
 import { IUserEdit } from '../../interfaces';
 import UserEditElement from './user/UserEditElement';
 
 interface UserEditProps {
   token: string | null;
-  setToken: (
-    value: string | ((val: string | null) => string | null) | null,
-  ) => void;
-  users: DataResponse<IUserEdit[]>;
+  setToken: (v: string | null) => void;
+  dataUsers: IUserEdit[] | undefined;
+  mutateUsers: KeyedMutator<IUserEdit[]>;
 }
 
-const UserEdit: FC<UserEditProps> = ({ token, setToken, users }) => {
+const UserEdit: FC<UserEditProps> = ({
+  token,
+  setToken,
+  dataUsers,
+  mutateUsers,
+}) => {
   return (
     <div className="mt-[30px]">
-      {users.data &&
-        users.data
+      {!!token &&
+        !!dataUsers &&
+        dataUsers
           .sort((a, b) => +a.id - +b.id)
-          .map((user) => (
+          .map((user, index) => (
             <UserEditElement
+              key={index}
               token={token}
               setToken={setToken}
               user={user}
-              mutate={users.mutate}
+              mutate={mutateUsers}
             />
           ))}
     </div>
