@@ -4,25 +4,16 @@ import axios from 'axios';
 import { FC, FormEvent, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { toast } from 'react-toastify';
-import { MutatorCallback } from 'swr/dist/types';
+import { KeyedMutator } from 'swr';
 import { ILectureInList } from '../../../interfaces';
 
 interface UpdateImageFieldProps {
   token: string | null;
-  setToken: (
-    value: string | ((val: string | null) => string | null) | null,
-  ) => void;
+  setToken: (v: string | null) => void;
   fieldType: string;
   lectureId: string;
   userField: string | null;
-  mutate: (
-    data?:
-      | ILectureInList[]
-      | Promise<ILectureInList[]>
-      | MutatorCallback<ILectureInList[]>
-      | undefined,
-    shouldRevalidate?: boolean | undefined,
-  ) => Promise<ILectureInList[] | undefined>;
+  mutate: KeyedMutator<ILectureInList[]>;
   imageIndex: number | null;
 }
 
